Corrected the listing to call the base Exception constructor
