+++ src/Chapter11/Listing11.06.DefiningASerializableException.cs
@@ -1,20 +1,71 @@
+using System;
+using System.Runtime.Serialization;
+
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter11.Listing11_06
 {
-   using System;
-   using System.Runtime.Serialization;
 
-   // Supporting serialization via an attribute
-   [Serializable]
-   class DatabaseException : System.ApplicationException
-   {
-       // ...
+    class DatabaseException : Exception
+    {
+        public DatabaseException(
+            string message,
+            System.Data.SqlClient.SQLException exception)
+            : base(message, innerException: exception)
+        {
+            // ...
+        }
+
+        public DatabaseException(
+            string message,
+            System.Data.OracleClient.OracleException exception)
+            : base(message, innerException: exception)
+        {
+            // ...
+        }
+
+        public DatabaseException()
+        {
+            // ...
+        }
+
+        public DatabaseException(string message)
+            : base(message)
+        {
+            // ...
+        }
+
+        public DatabaseException(
+            string message, Exception exception)
+            : base(message, innerException: exception)
+        {
+            // ...
+        }
 
        // Used for deserialization of exceptions
        public DatabaseException(
            SerializationInfo serializationInfo,
            StreamingContext context)
+           : base(serializationInfo, context)
        {
            //...
        }
-   }
+    }
+
+    // Create mock versions of the database exception classes rather
+    // than referencing the real libraries.
+    namespace System.Data
+    {
+        namespace SqlClient
+        {
+            class SQLException : Exception
+            {
+            }
+        }
+        namespace OracleClient
+        {
+            class OracleException : Exception
+            {
+            }
+        }
+    }
 }
+
